The BareBones Typscript with React - I
Simple with ReactDOm render that renders just a component. No Redux Integration. Nothing. Just run the tests and run the dev and you will see the 3 h1's coming on the browser when the webpack dev server is run. When you run the the tests, the barebones tests are run.